+++ nilearn/plotting/__init__.py
@@ -16,15 +16,17 @@ from nilearn.plotting.find_cuts import (
 from nilearn.plotting.html_connectome import view_connectome, view_markers
 from nilearn.plotting.html_stat_map import view_img
 from nilearn.plotting.html_surface import view_img_on_surf, view_surf
+from nilearn.plotting.img_comparison import (
+    plot_bland_altman,
+    plot_img_comparison,
+)
 from nilearn.plotting.img_plotting import (
     plot_anat,
-    plot_bland_altman,
     plot_carpet,
     plot_connectome,
     plot_epi,
     plot_glass_brain,
     plot_img,
-    plot_img_comparison,
     plot_markers,
     plot_prob_atlas,
     plot_roi,
